requirements.txt
email-validator==2.0.0.post2
fastapi==0.100.1
hubspot-api-client==8.0.0
psycopg2==2.9.6
requests==2.31.0
SQLAlchemy==2.0.19
uvicorn==0.23.2
